+++ src/modules/dns/unix/unabto_unix_dns.c
@@ -1,10 +1,11 @@
 #include <unabto/unabto_external_environment.h>
+#include <unabto/unabto_util.h>
 
 #include <pthread.h>
 
 typedef struct {
     const char* id;
-    uint32_t resolved_addr;
+    uint32_t resolved_addrs[NABTO_DNS_RESOLVED_IPS_MAX];
     nabto_dns_status_t status;
     pthread_t thread;
     bool thread_created;
@@ -20,9 +21,17 @@ void* resolver_thread(void* ctx) {
     struct hostent* he = gethostbyname(state->id);
     if (he == 0) {
         state->status = NABTO_DNS_ERROR;
-    } else {
+    } else if (he->h_addrtype == AF_INET && he->h_length == 4) {
+        uint8_t i;
         state->status = NABTO_DNS_OK;
-        state->resolved_addr = htonl(*((uint32_t*)he->h_addr_list[0]));
+
+        for (i = 0; i < NABTO_DNS_RESOLVED_IPS_MAX; i++) {
+            uint8_t* addr = (uint8_t*)he->h_addr_list[i];
+            if (addr == NULL) {
+                break;
+            }
+            READ_U32(state->resolved_addrs[i], addr);
+        }
     }
     resolver_is_running = false;
     return NULL;
@@ -48,8 +57,9 @@ static int create_detached_resolver() {
 
 void nabto_dns_resolve(const char* id) {
     uint32_t addr = inet_addr(id);
+    memset(resolver_state.resolved_addrs, 0, sizeof(uint32_t)*NABTO_DNS_RESOLVED_IPS_MAX); 
     if (addr != INADDR_NONE) {
-        resolver_state.resolved_addr = htonl(addr);
+        resolver_state.resolved_addrs[0] = htonl(addr);
         resolver_state.status = NABTO_DNS_OK;
     } else {
         // host isn't a dotted IP, so resolve it through DNS
@@ -68,13 +78,16 @@ void nabto_dns_resolve(const char* id) {
     }
 }
 
-nabto_dns_status_t nabto_dns_is_resolved(const char *id, uint32_t* v4addr) {
+nabto_dns_status_t nabto_dns_is_resolved(const char *id, uint32_t* v4addrs) {
     if (resolver_is_running) {
         return NABTO_DNS_NOT_FINISHED;
     }
     
     if (resolver_state.status == NABTO_DNS_OK) {
-        *v4addr = resolver_state.resolved_addr;
+        uint8_t i;
+        for (i = 0; i < NABTO_DNS_RESOLVED_IPS_MAX; i++) {
+            v4addrs[i] = resolver_state.resolved_addrs[i];
+        }
         return NABTO_DNS_OK;
     }
     return NABTO_DNS_ERROR;